OPENNLP-365 Re-factoring: boxing\unboxing, extra casts, shorter loops. Thanks to Aliaksandr Autayeu for providing a patch.
git-svn-id: https://svn.apache.org/repos/asf/incubator/opennlp/trunk@1199650 13f79535-47bb-0310-9956-ffa450edef68

+++ opennlp-tools/src/main/java/opennlp/tools/parser/chunking/BuildContextGenerator.java
@@ -57,7 +57,7 @@ public class BuildContextGenerator extends AbstractContextGenerator {
 
   public String[] getContext(Object o) {
     Object[] params = (Object[]) o;
-    return getContext((Parse[]) params[0], ((Integer) params[1]).intValue());
+    return getContext((Parse[]) params[0], (Integer) params[1]);
   }
 
   /**
